requirements.txt
python-telegram-bot>=21.0
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytesseract==0.3.10
Pillow>=10.0.0
dateparser==1.2.0
python-dotenv==1.0.1
